Make a copy of the geometry before modifying it
This bit of mutability hasn't actually caused any bugs so far (because
the stuff that needs to change, i.e. values and colours for the map, are
overwritten whenever necessary). However, it's nice to have an
unmodified, original copy of the geometries, so that we can re-export it
as part of the scenario zip file. (#53)