Refactor WASM scenario passing
Previously, the model identifier and the scenario were bundled together
as a JSON string and sent to the WASM worked. This commit disentangles
them so that they are sent as proper JS objects and only converted to
JSON internally when needed.

This commit also fixes issues in local dev mode arising from ambiguities
over the path to the wheel (when run in dev mode, the wheel is only
accessible from the root of the project, whereas when it's built, one
copy of the wheel is accessible in each /area_name subdirectory).